Separate between unit and integration tests
