+++ components/eventing-controller/pkg/backend/jetstreamv2/connection_integration_test.go
@@ -1,3 +1,6 @@
+//go:build integration
+// +build integration
+
 package jetstreamv2_test
 
 import (
