+++ components/eventing-controller/pkg/backend/jetstreamv2/jetstream_internal_unit_test.go
@@ -1,3 +1,6 @@
+//go:build unit
+// +build unit
+
 package jetstreamv2
 
 import (
