+++ components/eventing-controller/pkg/backend/jetstreamv2/utils_integration_test.go
@@ -1,0 +1,75 @@
+//go:build integration
+// +build integration
+
+package jetstreamv2
+
+import (
+	"testing"
+
+	eventingv1alpha2 "github.com/kyma-project/kyma/components/eventing-controller/api/v1alpha2"
+	evtestingv2 "github.com/kyma-project/kyma/components/eventing-controller/testing/v2"
+	"github.com/stretchr/testify/require"
+)
+
+// TestJetStream_isJsSubAssociatedWithKymaSub tests the isJsSubAssociatedWithKymaSub method.
+func TestJetStream_isJsSubAssociatedWithKymaSub(t *testing.T) {
+	t.Parallel()
+	// given
+	testEnvironment := setupTestEnvironment(t)
+	jsBackend := testEnvironment.jsBackend
+	defer testEnvironment.natsServer.Shutdown()
+	defer testEnvironment.jsClient.natsConn.Close()
+	initErr := jsBackend.Initialize(nil)
+	require.NoError(t, initErr)
+
+	// create subscription 1 and its JetStream subscription
+	cleanSubject1 := "subOne"
+	sub1 := evtestingv2.NewSubscription(cleanSubject1, "foo", evtestingv2.WithNotCleanEventSourceAndType())
+	jsSub1Key := NewSubscriptionSubjectIdentifier(sub1, cleanSubject1)
+
+	// create subscription 2 and its JetStream subscription
+	cleanSubject2 := "subOneTwo"
+	sub2 := evtestingv2.NewSubscription(cleanSubject2, "foo", evtestingv2.WithNotCleanEventSourceAndType())
+	jsSub2Key := NewSubscriptionSubjectIdentifier(sub2, cleanSubject2)
+
+	testCases := []struct {
+		name            string
+		givenJSSubKey   SubscriptionSubjectIdentifier
+		givenKymaSubKey *eventingv1alpha2.Subscription
+		wantResult      bool
+	}{
+		{
+			name:            "",
+			givenJSSubKey:   jsSub1Key,
+			givenKymaSubKey: sub1,
+			wantResult:      true,
+		},
+		{
+			name:            "",
+			givenJSSubKey:   jsSub2Key,
+			givenKymaSubKey: sub2,
+			wantResult:      true,
+		},
+		{
+			name:            "",
+			givenJSSubKey:   jsSub1Key,
+			givenKymaSubKey: sub2,
+			wantResult:      false,
+		},
+		{
+			name:            "",
+			givenJSSubKey:   jsSub2Key,
+			givenKymaSubKey: sub1,
+			wantResult:      false,
+		},
+	}
+
+	for _, tC := range testCases {
+		testCase := tC
+		t.Run(testCase.name, func(t *testing.T) {
+			t.Parallel()
+			gotResult := isJsSubAssociatedWithKymaSub(tC.givenJSSubKey, tC.givenKymaSubKey)
+			require.Equal(t, tC.wantResult, gotResult)
+		})
+	}
+}
